+++ poetry/console/application.py
@@ -20,6 +20,7 @@ from .commands import PublishCommand
 from .commands import RemoveCommand
 from .commands import RunCommand
 from .commands import ScriptCommand
+from .commands import SearchCommand
 from .commands import ShowCommand
 from .commands import UpdateCommand
 from .commands import VersionCommand
@@ -100,6 +101,7 @@ class Application(BaseApplication):
             RemoveCommand(),
             RunCommand(),
             ScriptCommand(),
+            SearchCommand(),
             ShowCommand(),
             UpdateCommand(),
             VersionCommand(),