+++ poetry/console/commands/__init__.py
@@ -10,6 +10,7 @@ from .publish import PublishCommand
 from .remove import RemoveCommand
 from .run import RunCommand
 from .script import ScriptCommand
+from .search import SearchCommand
 from .show import ShowCommand
 from .update import UpdateCommand
 from .version import VersionCommand