+++ poetry/repositories/pool.py
@@ -4,6 +4,7 @@ from typing import Union
 import poetry.packages
 
 from .base_repository import BaseRepository
+from .legacy_repository import LegacyRepository
 from .repository import Repository
 
 
@@ -79,4 +80,11 @@ class Pool(BaseRepository):
         return []
 
     def search(self, query, mode=BaseRepository.SEARCH_FULLTEXT):
-        raise NotImplementedError()
+        results = []
+        for repository in self._repositories:
+            if isinstance(repository, LegacyRepository):
+                continue
+
+            results += repository.search(query, mode=mode)
+
+        return results